@@ -15,11 +15,10 @@ setup(
         "pydantic-settings>=2.0.0",
         "fastapi>=0.100.0",
     ],
-    python_requires=">=3.10",
+    python_requires=">=3.11",
     classifiers=[
         "Development Status :: 4 - Beta",
         "Intended Audience :: Developers",
-        "Programming Language :: Python :: 3.10",
         "Programming Language :: Python :: 3.11",
     ],
 )
