@@ -4,19 +4,6 @@ Base models shared across all microservices
 from pydantic import BaseModel, Field
 from enum import Enum
 from datetime import datetime, timedelta
-try:
-    from common.datetime_utils import now_brazil
-except ImportError:
-    from datetime import timezone
-    try:
-        from zoneinfo import ZoneInfo
-    except ImportError:
-        from backports.zoneinfo import ZoneInfo
-    
-    BRAZIL_TZ = ZoneInfo("America/Sao_Paulo")
-    def now_brazil() -> datetime:
-        return datetime.now(BRAZIL_TZ)
-
 from typing import Optional
 import uuid
 
@@ -58,7 +45,7 @@ class BaseJob(BaseModel):
     started_at: Optional[datetime] = None
     completed_at: Optional[datetime] = None
     expires_at: datetime = Field(
-        default_factory=lambda: now_brazil() + timedelta(hours=24)
+        default_factory=lambda: datetime.now() + timedelta(hours=24)
     )
     
     # Progress tracking
@@ -84,7 +71,7 @@ class BaseJob(BaseModel):
     @property
     def is_expired(self) -> bool:
         """Verifica se o job expirou"""
-        return now_brazil() > self.expires_at
+        return datetime.now() > self.expires_at
     
     @property
     def is_terminal(self) -> bool:
@@ -100,7 +87,7 @@ class BaseJob(BaseModel):
         """Retorna duração do job em segundos"""
         if not self.started_at:
             return None
-        end_time = self.completed_at or now_brazil()
+        end_time = self.completed_at or datetime.now()
         return (end_time - self.started_at).total_seconds()
     
     def mark_as_queued(self):
@@ -111,14 +98,14 @@ class BaseJob(BaseModel):
         """Marca job como em processamento"""
         self.status = JobStatus.PROCESSING
         if not self.started_at:
-            self.started_at = now_brazil()
+            self.started_at = datetime.now()
         if message:
             self.progress_message = message
     
     def mark_as_completed(self, message: Optional[str] = None):
         """Marca job como completado"""
         self.status = JobStatus.COMPLETED
-        self.completed_at = now_brazil()
+        self.completed_at = datetime.now()
         self.progress = 100.0
         if message:
             self.progress_message = message
@@ -126,14 +113,14 @@ class BaseJob(BaseModel):
     def mark_as_failed(self, error: str, error_type: Optional[str] = None):
         """Marca job como falho"""
         self.status = JobStatus.FAILED
-        self.completed_at = now_brazil()
+        self.completed_at = datetime.now()
         self.error_message = error
         self.error_type = error_type or "UnknownError"
     
     def mark_as_cancelled(self, reason: Optional[str] = None):
         """Marca job como cancelado"""
         self.status = JobStatus.CANCELLED
-        self.completed_at = now_brazil()
+        self.completed_at = datetime.now()
         if reason:
             self.error_message = f"Cancelled: {reason}"
     
